Full 0.4.5 update
This version is basically the same that was published, but fully updated to Julia 0.4.5, plots and all. Also updated some minor implementations to plots (auto axis range), and fixed some minor bugs).

The main program (MCRAFT) now returns a vector of vectors. And this whole vector is used to plot.